Create bios.yaml
The other day, I had to write a bio. I realized that I've never made a place to keep track of bios I've had for myself. 

I want to keep track now because they are hard to remember, not always findable, and a great way to measure personal development over time.